Re-factored the Docker functionality for mapping Dockerfile
The Dockerfile wasn't being included in the `mappings` setting, but it now is. Along the way I took the opportunity of cleaning up how stage works, bringing it more inline with the approach taken by the Universal plugin.

One change that occurred as a result of the above is that `target/docker` is now used to generate working files (`Dockerfile`) as well as holding a `stage` folder. The `stage` folder contains the `Dockerfile` for staging add its context files. This is very similar to the structure of the `target/universal` folder.