If the /var/run (aka /run) folder doesn't exist for the PID file, create it
CentOS 7 and many other recent Linuxes have switched to a non-persistent (tmpfs) /run folder, with /var/run being a symbolic link to the tmpfs /run mount.  The RPM creates the correct folder in /var/run for the PID file, but on reboot, that file is gone.  There are many ways to fix it.  This way will work on older systemd versions.