Added CombinedCmd and removed EnvCmd
I added CombinedCmd for ONBUILD instruction and possible future commands that will be added that will have a similar form. For readability I made it so that it uses a CmdLike object as an argument. Cmd could also be used with an extra argument but it would degrade readability. Removed ENV because it introduced an inconsistency with ADD and COPY. Made `Cmd` accept multiple arguments because this solves the ADD and ENV issue. This fix won't break the old api because a string with a space in the middle is still acceptable only less readable.