"docker push" should support dockerExecCommand (#911)
* When building docker images on CentOS, Fedora or RHEL, the permissions on /var/run/docker.socket have restricted access, preventing Docker build and push commands from executing without severely compromising the system.

See http://www.projectatomic.io/blog/2015/08/why-we-dont-let-non-root-users-run-docker-in-centos-fedora-or-rhel/

Therefore, I've added a new setting which supports replacing the default docker shell exec command so that I can instead use `sudo docker`, which is the recommended method for executing Docker on CentOS, Fedora and RHEL.

What I don't want to do is run SBT as root - this will cause the working directory, and the ivy cache to become polluted with artifacts that are owned by root. It's best to limit the use of escalated privileges to just those commands that require them to avoid this pollution.

So, to avoid:

sudo sbt docker:publishLocal
sbt clean <=== will fail because "target" is now owned by root

* dockerExecCommand should be used for "docker push"

I was certain that I made this change initially - perhaps it got lost
during some of my rebasing.

The "docker push" should also support the dockerExecCommand setting.