Add Setting for Docker User
Currently, there is no way to set the organizaiton or user for a docker
image. The documented practice is to misuse the dockerRepository setting
to set the username or organization. However, there is already
functionality in place for the DockerAlias class to support a separate
username.

This commit adds a dockerUsername setting which ties into the existing
DockerAlias username functionality.